[overrides] Deprecate "Delete all overrides" and add "Delete"
In the Sources panel > Overrides tab, right click a folder to see its context menu.

- Fixed "Delete all overrides" to includes .header file during delete
- Deprecate "Delete all overrides", only enable it if users enable the new Experiment > Enable "Delete all overrides" temporarily
- New "Delete" menu to delete the selected folder and its content
- A stronger warning message in the delete confirmation
- The new "Delete" is a real "Delete" - Not only delete all the files within a selected folder, but also delete the folders within it.
- Add metric to track on usage, so we can decide later on the deprecation

screenshot: https://crbug.com/1473681

Bug: 1416338,1472580,1473681
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4777899
Commit-Queue: Jecelyn Yeen <jec@chromium.org>

+++ front_end/panels/sources/NavigatorView.ts
@@ -101,7 +101,15 @@ const UIStrings = {
   /**
    *@description Text in Navigator View of the Sources panel
    */
-  areYouSureYouWantToDeleteAll: 'Are you sure you want to delete all overrides in this folder?',
+  areYouSureYouWantToDeleteAllOverrides: 'Are you sure you want to delete all overrides in this folder?',
+  /**
+   *@description Text in Navigator View of the Sources panel
+   */
+  areYouSureYouWantToDeleteFolder: 'Are you sure you want to delete this folder and its contents?',
+  /**
+   *@description Text in Navigator View of the Sources panel. A confirmation message on action to delete a folder.
+   */
+  actionCannotBeUndone: 'This action cannot be undone.',
   /**
    *@description A context menu item in the Navigator View of the Sources panel
    */
@@ -996,8 +1004,10 @@ export class NavigatorView extends UI.Widget.VBox implements SDK.TargetManager.O
   }
 
   private async handleDeleteOverrides(node: NavigatorTreeNode): Promise<void> {
-    const shouldRemove = await UI.UIUtils.ConfirmDialog.show(i18nString(UIStrings.areYouSureYouWantToDeleteAll));
+    const shouldRemove =
+        await UI.UIUtils.ConfirmDialog.show(i18nString(UIStrings.areYouSureYouWantToDeleteAllOverrides));
     if (shouldRemove) {
+      Host.userMetrics.actionTaken(Host.UserMetrics.Action.OverrideTabDeleteOverridesContextMenu);
       this.handleDeleteOverridesHelper(node);
     }
   }
@@ -1006,13 +1016,65 @@ export class NavigatorView extends UI.Widget.VBox implements SDK.TargetManager.O
     node.children().forEach(child => {
       this.handleDeleteOverridesHelper(child);
     });
+
     if (node instanceof NavigatorUISourceCodeTreeNode) {
       // Only delete confirmed overrides and not just any file that happens to be in the folder.
       const binding = Persistence.Persistence.PersistenceImpl.instance().binding(node.uiSourceCode());
-      if (binding) {
+      const headerBinding =
+          Persistence.NetworkPersistenceManager.NetworkPersistenceManager.instance().isActiveHeaderOverrides(
+              node.uiSourceCode());
+      if (binding || headerBinding) {
         node.uiSourceCode().project().deleteFile(node.uiSourceCode());
       }
     }
+  }
+
+  private async handleDeleteFolder(node: NavigatorTreeNode): Promise<void> {
+    const warningMsg =
+        `${i18nString(UIStrings.areYouSureYouWantToDeleteFolder)}\n${i18nString(UIStrings.actionCannotBeUndone)}`;
+    const shouldRemove = await UI.UIUtils.ConfirmDialog.show(warningMsg);
+    if (shouldRemove) {
+      Host.userMetrics.actionTaken(Host.UserMetrics.Action.OverrideTabDeleteFolderContextMenu);
+      const topNode = this.findTopNonMergedNode(node);
+      await this.removeUISourceCodeFromProject(topNode);
+      await this.deleteDirectoryRecursively(topNode);
+    }
+  }
+
+  private async removeUISourceCodeFromProject(node: NavigatorTreeNode): Promise<void> {
+    node.children().forEach(async child => {
+      await this.removeUISourceCodeFromProject(child);
+    });
+
+    if (node instanceof NavigatorUISourceCodeTreeNode) {
+      node.uiSourceCode().project().removeUISourceCode(node.uiSourceCode().url());
+    }
+  }
+
+  private async deleteDirectoryRecursively(node: NavigatorTreeNode): Promise<void> {
+    if (!(node instanceof NavigatorFolderTreeNode)) {
+      return;
+    }
+
+    await Persistence.NetworkPersistenceManager.NetworkPersistenceManager.instance()
+        .project()
+        ?.deleteDirectoryRecursively(node.folderPath);
+  }
+
+  private findTopNonMergedNode(node: NavigatorTreeNode): NavigatorTreeNode {
+    // multiple folder nodes can be merged into one if it only contains one file
+    // e.g. the folder of "abc.com/assets/css/button.css" can be "abc.com/assets/css"
+    // find the top non-merged node (abc.com) recursively
+
+    if (!node.isMerged) {
+      return node;
+    }
+
+    if (!(node.parent instanceof NavigatorFolderTreeNode)) {
+      return node;
+    }
+
+    return this.findTopNonMergedNode(node.parent);
   }
 
   handleFolderContextMenu(event: Event, node: NavigatorFolderTreeNode): void {
@@ -1074,8 +1136,15 @@ export class NavigatorView extends UI.Widget.VBox implements SDK.TargetManager.O
           });
         }
       } else {
-        contextMenu.defaultSection().appendItem(
-            i18nString(UIStrings.deleteAllOverrides), this.handleDeleteOverrides.bind(this, node));
+        if (Root.Runtime.experiments.isEnabled(Root.Runtime.ExperimentName.DELETE_OVERRIDES_TEMP_ENABLE)) {
+          contextMenu.defaultSection().appendItem(
+              i18nString(UIStrings.deleteAllOverrides), this.handleDeleteOverrides.bind(this, node));
+        }
+
+        if (!(node instanceof NavigatorGroupTreeNode)) {
+          contextMenu.defaultSection().appendItem(
+              i18nString(UIStrings.delete), this.handleDeleteFolder.bind(this, node));
+        }
       }
     }
 
@@ -1333,19 +1402,16 @@ export class NavigatorSourceTreeElement extends UI.TreeOutline.TreeElement {
 
   updateIcon(): void {
     const binding = Persistence.Persistence.PersistenceImpl.instance().binding(this.uiSourceCodeInternal);
+    const networkPersistenceManager = Persistence.NetworkPersistenceManager.NetworkPersistenceManager.instance();
     let iconType = 'document';
     let iconStyles: string[] = [];
     if (binding) {
       if (Snippets.ScriptSnippetFileSystem.isSnippetsUISourceCode(binding.fileSystem)) {
         iconType = 'snippet';
       }
-      const badgeIsPurple = Persistence.NetworkPersistenceManager.NetworkPersistenceManager.instance().project() ===
-          binding.fileSystem.project();
+      const badgeIsPurple = networkPersistenceManager.project() === binding.fileSystem.project();
       iconStyles = badgeIsPurple ? ['dot', 'purple'] : ['dot', 'green'];
-    } else if (
-        this.uiSourceCodeInternal.url().endsWith(Persistence.NetworkPersistenceManager.HEADERS_FILENAME) &&
-        Persistence.NetworkPersistenceManager.NetworkPersistenceManager.instance()
-            .hasMatchingNetworkUISourceCodeForHeaderOverridesFile(this.uiSourceCodeInternal)) {
+    } else if (networkPersistenceManager.isActiveHeaderOverrides(this.uiSourceCode)) {
       iconStyles = ['dot', 'purple'];
     } else {
       if (Snippets.ScriptSnippetFileSystem.isSnippetsUISourceCode(this.uiSourceCodeInternal)) {
